+++ DesktopApplicationTemplate.UI/Views/TcpServiceMessagesView.xaml.cs
@@ -1,4 +1,5 @@
 using System.Linq;
+using System.Windows;
 using System.Windows.Controls;
 using DesktopApplicationTemplate.UI.ViewModels;
 
@@ -20,6 +21,26 @@ namespace DesktopApplicationTemplate.UI.Views
             LogView.DataContext = new ServiceLogViewModel(service.DisplayName, service.ServiceType, service.Logs);
             if (DataContext is TcpServiceMessagesViewModel vm)
                 vm.SetService(service);
+        }
+
+        private void EditScript_Click(object sender, RoutedEventArgs e)
+        {
+            if (DataContext is not TcpServiceMessagesViewModel vm)
+                return;
+
+            var editor = new ScriptEditorWindow();
+            if (editor.DataContext is ScriptEditorViewModel svm)
+            {
+                svm.ScriptText = vm.Script;
+                svm.TestMessage = vm.TestMessage;
+            }
+
+            if (editor.ShowDialog() == true)
+            {
+                vm.Script = editor.ScriptText;
+                vm.TestMessage = editor.LastTestMessage;
+                vm.Save();
+            }
         }
     }
 }